BatchPM {GCDmodel}	R Documentation

Trace elements concentrations for a batch process

Description

More commonly used for batch melting, but does actually also work with crystalization. Make sure to use the same mineral/element names in all the variables

Usage

BatchPM(kd, c0, pm, cmins = matrix(), min.props, melt.arg = list(),
  dont = character(0))

Arguments

kd
Named matrix. Partition coefficients

c0
Named vector. The composition of the source

pm
Scalar. Degree of partial melting (0-100)

cmins
Matrix. Composition of minerals (not used, for consistency)

min.props
Named vector. Mineral proportions (normalized to 1)

melt.arg
List. Further useful arguments (mostly used for correction modules, kept here for consistency)

dont
Character vectors. Do not calculate for these elements. Not used here, kept for consistency (mostly used in correction modules)

Value

list containing the following slots:

cL Named vector. Composition of the liquid

cmins Named matrix. Composition of all individual solid mineral phases

cS Named vector. Composition of the whole solid

kd Named matrix. The Kd table, here similar to the input (included for consistency with other modelling routines)

DD Named vector. The D (bulk distribution) values.

[Package GCDmodel version 0.4.7 Index]